Chronos/requirements.txt
streamlit>=1.35
google-genai>=0.8
pillow>=10.0
python-dotenv>=1.0
